functions_ingest_sample/requirements.txt

azure-functions
requests
